app/backend/requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.13.1,<3.14
pyjwt>=2.4.0,<2.5.0
drf-spectacular>=0.23.1,<0.24
psycopg2>=2.9.3,<2.10
#psycopg2-binary #dev
